Add return deleted subscriber

+++ app/views/subscriber/index.blade.php
@@ -52,12 +52,23 @@
                             echo implode(",",$gr_list);
                         ?>
                     </td>
-                    <td>{{$subscriber->created_at}}</td>
+                    <td>
+                        {{$subscriber->created_at}}
+                        @if empty($subscriber->deleted_at)
+                            удален: {{$subscriber->deleted_at}}
+                        @endif
+                    </td>
                     <td>
                         <?/*<a href="{{URL::action('SubscriberController@edit',array('id'=>$subscriber->id))}}" class="btn btn-default btn-mini"><i class="fa fa-edit"></i></a>*/   ?>
-                        {{Form::open(array('url' => URL::action('SubscriberController@destroy',array('id'=>$subscriber->id)), 'method' => 'delete','style'=>'display:inline-block')) }}
-                        <button onClick="if(!confirm('{{trans('subscriber.messageDelete')}}')) return false;" type="submit" class="btn btn-danger btn-mini"><i class="fa fa-trash-o"></i></button>
-                        {{Form::close();}}
+                        @if empty($subscriber->deleted_at)
+                            {{Form::open(array('url' => URL::action('SubscriberController@destroy',array('id'=>$subscriber->id)), 'method' => 'delete','style'=>'display:inline-block')) }}
+                            <button onClick="if(!confirm('{{trans('subscriber.messageDelete')}}')) return false;" type="submit" class="btn btn-danger btn-mini"><i class="fa fa-trash-o"></i></button>
+                            {{Form::close();}}
+                        @else
+                            {{Form::open(array('url' => URL::action('SubscriberController@update',array('id'=>$subscriber->id)), 'method' => 'delete','style'=>'display:inline-block')) }}
+                            <button type="submit" class="btn btn-success btn-mini"><i class="fa fa-back-o"></i></button>
+                            {{Form::close();}}
+                        @endif
                     </td>
                 </tr>
                 @endforeach
